remove: footer.js로 통합
